requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
httpie==3.2.3
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
pillow==10.4.0
pygments==2.18.0
PySocks==1.7.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.1
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
